move `benchmarks/util.py` to `tests/util/benchmarks.py`
